+++ lib/sparse_autoencoders/pyproject.toml
@@ -1,0 +1,38 @@
+[project]
+name = "sparse_autoencoders"
+version = "0.0.1"
+description = "Sparse autoencoders"
+
+requires-python = "==3.12.*"
+dependencies = [
+  "activations",
+  "baukit",
+  "einops>=0.8.0",
+  "jacobian-saes",
+  "sae-lens",
+  "sae",
+  "transformer-lens>=2.11.0",
+  "util",
+]
+
+[tool.uv.sources]
+activations = { path = "../activations", editable = true }
+baukit = { git = "https://github.com/choidami/baukit.git" }
+jacobian-saes = { path = "../../../jacobian-saes", editable = true }
+sae = { git = "https://github.com/EleutherAI/sae" }
+sae-lens = { git = "https://github.com/tim-lawson/SAELens", branch = "timl/deps" }
+util = { path = "../util", editable = true }
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["sparse_autoencoders"]
+
+[tool.pyright]
+strict = ["sparse_autoencoders"]
+reportMissingTypeStubs = false
+
+[tool.hatch.metadata]
+allow-direct-references = true
