+++ lib/sparse_autoencoders/sparse_autoencoders/util.py
@@ -1,0 +1,13 @@
+import torch
+
+
+def get_torch_dtype(dtype: str) -> torch.dtype:
+    if dtype == "bfloat16":
+        return torch.bfloat16
+    if dtype == "float16":
+        return torch.float16
+    if dtype == "float32":
+        return torch.float32
+    if dtype == "float64":
+        return torch.float64
+    raise ValueError(f"Unknown dtype: {dtype}")
