+++ project/expgen/.gitignore
@@ -1,1 +1,2 @@
 data/
+temp/
